Changing type of volume variable

+++ Entries/VolumeEntry.cs
@@ -7,7 +7,7 @@ namespace FantomLis.BoomboxExtended.Entries;
 public class VolumeEntry : BaseEntry, IHaveUIData
 {
     public VolumeEntry() : this (50){}
-    public int Volume { get; private set; }
+    public float Volume { get; private set; }
 
     private string VolumeText = $"{{0}}% {BoomboxLocalization.BoomboxVolume}";
 
@@ -15,7 +15,7 @@ public class VolumeEntry : BaseEntry, IHaveUIData
     /// Updates volume
     /// </summary>
     /// <param name="vol">Volume from 0 to 100</param>
-    public void UpdateVolume(int vol)
+    public void UpdateVolume(float vol)
     {
         Volume = Math.Clamp(vol, 0, 100);
         SetDirty();
@@ -25,26 +25,26 @@ public class VolumeEntry : BaseEntry, IHaveUIData
     /// Decrease volume
     /// </summary>
     /// <param name="vol">How much decrease volume</param>
-    public void MinusVolume(int vol) => UpdateVolume(Volume - vol);
+    public void MinusVolume(float vol) => UpdateVolume(Volume - vol);
     /// <summary>
     /// Increase volume
     /// </summary>
     /// <param name="vol">How much increase volume</param>
-    public void PlusVolume(int vol) => UpdateVolume(Volume + vol);
+    public void PlusVolume(float vol) => UpdateVolume(Volume + vol);
 
-    public VolumeEntry(int vol = 50)
+    public VolumeEntry(float vol = 50)
     {
         Volume = Math.Clamp(vol, 0, 100);
     }
 
     public override void Deserialize(BinaryDeserializer binaryDeserializer)
     {
-        Volume = binaryDeserializer.ReadInt();
+        Volume = binaryDeserializer.ReadFloat();
     }
 
     public override void Serialize(BinarySerializer binarySerializer)
     {
-        binarySerializer.WriteInt(Volume);
+        binarySerializer.WriteFloat(Volume);
     }
 
     public float GetVolume() => Volume / 100f;
